backend/requirements.txt
crewai==0.80.0
crewai-tools==0.17.0
anthropic==0.40.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.12
apscheduler==3.10.4
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
rich==13.9.4
# Resume parsing
pdfminer.six==20221105
pypdf==4.3.1
python-docx==1.1.2
